% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sgRNA_reference.R
\name{create_sgRNA_reference}
\alias{create_sgRNA_reference}
\title{A function to create a reference fasta and gtf for plasmid + gRNA fusion for CROP-seq screens}
\usage{
create_sgRNA_reference(sgRNA_table, name = "name",
  sequence = "sgRNA_sequence", assay = "original", outdir = ".")
}
\arguments{
\item{sgRNA_table}{The table containing the sgRNA sequences and annotation.}

\item{name}{The name of the column in sgRNA_table that should be used to label the fasta sequences}

\item{sequence}{The actual sgRNA sequence}

\item{assay}{The type of CRISPR assay that is being performed. If set to "activation", the tracr backbone is replaced by a tracr MS2 backbone used in VP64 screens.}

\item{outdir}{The directory where the .fasta and .gtf files will be saved. Default = current working directory (".")}
}
\description{
This function takes as input a table containing the gRNA sequences and will output a .fasta and a .gtf file \
containing the merged plasmid + gRNA sequences and the gtf description to add them to a reference genome.
}
\examples{
create_sgRNA_reference()
}
\keyword{Crop-seq,}
\keyword{reference}
\keyword{sgRNA,}
