% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DEseq2_guides.R
\name{run_deseq2_guides}
\alias{run_deseq2_guides}
\title{A function to calculate DESeq2 log-fold changes for sgRNA counts from poole CRISPR screens}
\usage{
run_deseq2_guides(design_matrix, sgRNA_count_table,
  design_column = "distribution")
}
\arguments{
\item{sgRNA_table}{The table containing the sgRNA sequences and annotation.}
}
\description{
This function takes as input a table containing the gRNA sequences and will output a .fasta and a .gtf file \
containing the merged plasmid + gRNA sequences and the gtf description to add them to a reference genome.
}
\examples{
run_deseq2_on_guides()
}
